+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -20,6 +20,7 @@
 
 ## Upcoming Release
 * Updated Azure.Core from 1.45.0 to 1.47.3
+* Source list parameter for redistribute throughput optional as it is no longer required
 
 ## Version 2.0.8
 * Fixed issue where the restore command failed to detect the source account, even when it existed.