+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 
 ## Upcoming Release
+* Updated Azure.Core from 1.45.0 to 1.47.3
+
+## Version 1.18.2
+* Fixed issue where the restore command failed to detect the source account, even when it existed.
 
 ## Version 1.18.1
 * Fixed incorrect help documents for MongoDB cmdlets.